requirements.txt
google-cloud-secret-manager
protobuf==4.25.2
tweepy==4.14.0
